Explicitly instantiate templated ValueRefParsers
This reduces overall compile time by only compiling the complex parsers
once.
