+++ parse/ConditionParser7.cpp
@@ -3,6 +3,7 @@
 #include "ValueRefParserImpl.h"
 
 #include "../universe/Condition.h"
+#include "../universe/ValueRef.h"
 
 #include <boost/spirit/include/phoenix.hpp>
