+++ parse/EffectParser4.cpp
@@ -2,6 +2,7 @@
 
 #include "ValueRefParserImpl.h"
 #include "../universe/Effect.h"
+#include "../universe/ValueRef.h"
 
 #include <boost/spirit/include/phoenix.hpp>
