+++ parse/IntValueRefParser.cpp
@@ -1,5 +1,9 @@
 #include "ValueRefParserImpl.h"
 
+#include "MovableEnvelope.h"
+#include "../universe/ValueRef.h"
+
+#include <boost/spirit/include/phoenix.hpp>
 
 parse::detail::simple_int_parser_rules::simple_int_parser_rules(const parse::lexer& tok) :
     simple_variable_rules("integer", tok)
