+++ parse/StringValueRefParser.cpp
@@ -1,6 +1,9 @@
 #include "ValueRefParserImpl.h"
 
+#include "MovableEnvelope.h"
 #include "../universe/ValueRef.h"
+
+#include <boost/spirit/include/phoenix.hpp>
 
 namespace parse {
     string_parser_grammar::string_parser_grammar(
@@ -57,11 +60,11 @@ namespace parse {
             |   tok.GalaxySeed_     [ _val = construct_movable_(new_<ValueRef::Variable<std::string>>(ValueRef::NON_OBJECT_REFERENCE, _1)) ]
             ;
 
-        variable_scope_rule = variable_scope(tok);
-        container_type_rule = container_type(tok);
-        initialize_bound_variable_parser<std::string>(bound_variable, bound_variable_name,
-                                                      variable_scope_rule, container_type_rule,
-                                                      tok);
+        variable_scope_rule = detail::variable_scope(tok);
+        container_type_rule = detail::container_type(tok);
+        detail::initialize_bound_variable_parser<std::string>(bound_variable, bound_variable_name,
+                                                              variable_scope_rule, container_type_rule,
+                                                              tok);
 
         statistic_sub_value_ref
             =   constant
@@ -116,7 +119,7 @@ namespace parse {
             =   operated_expr
             ;
 
-        initialize_nonnumeric_statistic_parser<std::string>(
+        detail::initialize_nonnumeric_statistic_parser<std::string>(
             statistic, tok, labeller, condition_parser, statistic_sub_value_ref);
 
         primary_expr
