+++ parse/VisibilityValueRefParser.cpp
@@ -5,6 +5,8 @@
 #include "../universe/Enums.h"
 #include "../universe/ValueRef.h"
 
+#include <boost/spirit/include/phoenix.hpp>
+
 namespace parse { namespace detail {
     visibility_complex_parser_grammar::visibility_complex_parser_grammar(
         const parse::lexer& tok, Labeller& labeller
